+++ app/src/androidTest/kotlin/com/mgaetan89/showsrage/extension/realm/RealmExtension_GetHistoryTest.kt
@@ -1,5 +1,6 @@
 package com.mgaetan89.showsrage.extension.realm
 
+import android.os.Looper
 import android.support.test.annotation.UiThreadTest
 import android.support.test.runner.AndroidJUnit4
 import com.mgaetan89.showsrage.buildComparator
@@ -7,6 +8,8 @@ import com.mgaetan89.showsrage.extension.getHistory
 import com.mgaetan89.showsrage.model.History
 import io.realm.RealmChangeListener
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.AfterClass
+import org.junit.BeforeClass
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -21,5 +24,21 @@ class RealmExtension_GetHistoryTest : RealmTest() {
             assertThat(it).hasSize(100)
             assertThat(it).isSortedAccordingTo(buildComparator(History::date, true))
         })
+    }
+
+    companion object {
+        @BeforeClass
+        @JvmStatic
+        fun beforeClass() {
+            if (Looper.myLooper() == null) {
+                Looper.prepare()
+            }
+        }
+
+        @AfterClass
+        @JvmStatic
+        fun afterClass() {
+            Looper.myLooper().quit()
+        }
     }
 }
