+++ app/src/androidTest/kotlin/com/mgaetan89/showsrage/extension/realm/RealmExtension_GetScheduleTest.kt
@@ -1,10 +1,13 @@
 package com.mgaetan89.showsrage.extension.realm
 
+import android.os.Looper
 import android.support.test.annotation.UiThreadTest
 import android.support.test.runner.AndroidJUnit4
 import com.mgaetan89.showsrage.extension.getSchedule
 import io.realm.RealmChangeListener
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.AfterClass
+import org.junit.BeforeClass
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -32,5 +35,21 @@ class RealmExtension_GetScheduleTest : RealmTest() {
 
             assertThat(it).isEmpty()
         })
+    }
+
+    companion object {
+        @BeforeClass
+        @JvmStatic
+        fun beforeClass() {
+            if (Looper.myLooper() == null) {
+                Looper.prepare()
+            }
+        }
+
+        @AfterClass
+        @JvmStatic
+        fun afterClass() {
+            Looper.myLooper().quit()
+        }
     }
 }
