+++ app/src/androidTest/kotlin/com/mgaetan89/showsrage/extension/realm/RealmTest.kt
@@ -1,6 +1,5 @@
 package com.mgaetan89.showsrage.extension.realm
 
-import android.os.Looper
 import android.support.test.InstrumentationRegistry
 import android.support.test.rule.ActivityTestRule
 import com.mgaetan89.showsrage.Constants
@@ -8,12 +7,9 @@ import com.mgaetan89.showsrage.TestActivity
 import com.mgaetan89.showsrage.helper.Migration
 import io.realm.Realm
 import io.realm.RealmConfiguration
-import io.realm.clearContext
 import io.realm.setContext
 import org.junit.After
-import org.junit.AfterClass
 import org.junit.Before
-import org.junit.BeforeClass
 import org.junit.Rule
 
 abstract class RealmTest {
@@ -25,7 +21,7 @@ abstract class RealmTest {
 
     @Before
     fun configureRealm() {
-        clearContext()
+        setContext(null)
 
         Realm.init(this.activityRule.activity)
 
@@ -48,22 +44,6 @@ abstract class RealmTest {
     fun after() {
         this.realm.close()
 
-        clearContext()
-    }
-
-    companion object {
-        @BeforeClass
-        @JvmStatic
-        fun beforeClass() {
-            if (Looper.myLooper() == null) {
-                Looper.prepare()
-            }
-        }
-
-        @AfterClass
-        @JvmStatic
-        fun afterClass() {
-            Looper.myLooper().quit()
-        }
+        setContext(null)
     }
 }
