+++ app/src/androidTest/kotlin/io/realm/RealmHelper.kt
@@ -2,10 +2,6 @@ package io.realm
 
 import android.content.Context
 
-fun clearContext() {
-    BaseRealm.applicationContext = null
-}
-
-fun setContext(context: Context) {
+fun setContext(context: Context?) {
     BaseRealm.applicationContext = context
 }
